requirements.txt
Django==2.2
djangorestframework==3.9.2
django-crontab
gunicorn==20.0.4
psycopg2
psycopg2-binary==2.9.1